+++ competition-serv-model/src/main/java/compman/compsrv/model/dto/brackets/StageDescriptorDTO.java
@@ -5,8 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -26,5 +24,5 @@ public class StageDescriptorDTO {
         private Boolean hasThirdPlaceFight;
         private GroupDescriptorDTO[] groupDescriptors;
         private Integer numberOfFights;
-        private BigDecimal fightDuration;
+        private Integer fightDuration;
 }
